Some refactoring (Dumped local SortDirection etc)

+++ CryptoScanner.Core/Model/CryptoData.cs
@@ -102,7 +102,7 @@ public class CryptoData
     public double? PSar { get; set; }
 
     [Computed]
-    public float Lux5mValue { get; set; }
+    public int Lux5mValue { get; set; }
 
     /// <summary>
     /// Copy common indicator values
